refactor(voyage): engager un trajet dans le parcours devient appelable depuis un module
`pickJourney` est le point d'entrée du compagnon de voyage, et les QUATRE vues
qui restent dans `app.js` l'appellent — Trajets en mode simple ET en mode
multi, Boucles, Chaîne — plus le manifeste d'« En route » et le déplacement
d'arrêt. Aucune ne peut emménager dans l'arbre tant qu'il vit dans un fichier
qui n'exporte rien.

Il rejoint `syncViewsToJourney` dans un `voyage-actions.ts` distinct de
`voyage-donnees.ts` : celui-ci porte la DONNÉE du parcours et s'interdit de
toucher au cycle de rendu, celui-là appelle `rafraichir()`.

`rafraichir()` remplace le couple `renderJourney(); refresh();` : `refresh()`
rejoue déjà la carte du voyage dès que `etat.JOURNEY` existe — et il existe
forcément là, `addToJourney` vient de le poser. Le premier appel ne faisait que
peindre la carte un tour plus tôt, dans la même tâche : rien ne pouvait
l'observer entre les deux.

`syncViewsToJourney` gagne une garde sur les champs absents : elle écrivait
`$("origin").value` sans vérifier que le nœud existe.

`addToJourney` et `currentLeg` quittent les imports d'app.js.

Vérification : 236 e2e (234 passés, 2 ignorés, 0 échec) · 486/486 unitaires ·
tsc silencieux · build OK.
